refactor: name the promotion reconciler for what it is
It has no ref, computed, watch or lifecycle hook — it is a plain factory
with an explicit dispose(), so the use* prefix promises reactivity it does
not have, and src/composables/ is the wrong home for graph-domain
machinery. Moves it next to resolveConcretePromotedWidget, which it pairs
with conceptually.
